Implement immediate generation in Pipeline

+++ src/binutils.h
@@ -35,14 +35,14 @@ constexpr uint32_t bitcount(int n) {
 
 inline uint32_t accBVec(std::vector<bool> v) {
     uint32_t r = 0;
-    for (size_t i = 0; i < v.size(); i++) {
+    for (uint32_t i = 0; i < v.size(); i++) {
         r |= v[i] << i;
     }
     return r;
 }
 
 inline void buildVec(std::vector<bool>& v, uint32_t n) {
-    for (size_t i = 0; i < v.size(); i++) {
+    for (uint32_t i = 0; i < v.size(); i++) {
         v[i] = n & 0b1;
         n >>= 1;
     }
